Mini-Projects/requirements.txt
pyfiglet
requests
bs4
pywhatkit
wikipedia
beautifulsoup4
requests
pyautogui
speedtest
pillow
pynput
speechrecognition
tqdm
